* chapter 10. "Metal"

+++ kernels/ray.cuh
@@ -44,7 +44,7 @@ public:
   {
     return dir;
   }
-  __device__ point3 at(double t) const
+  __device__ point3 at(float t) const
   {
     return orig + t*dir;
   }
